+++ mlib/randomize.cpp
@@ -1,22 +1,22 @@
 
-
 #include "mlib/randomize.hpp"
 
 namespace mlib
 {
-    void Rand::Randomize()
-    {
-        srand ( ( unsigned ) time ( NULL ) );
-    }
+	void Rand::Randomize()
+	{
+		srand ( ( size_t ) time ( NULL ) );
+	}
 
-    int Rand::Next ()
-    {
-        return rand();
-    }
+	size_t Rand::Next ()
+	{
+		return rand();
+	}
 
-    int Rand::Next(unsigned int from, unsigned int to)
-    {
-        return ((rand()) % (to-from) + from);
-    }
+	size_t Rand::Next(size_t from, size_t to)
+	{
+		return ((rand()) % (to-from) + from);
+	}
 
 }
+
